backend/requirements.txt
fitz
pymupdf
langchain
langchain_community
langchain_groq
markdown-it-py
fastapi
uvicorn 
psycopg2-binary
python-multipart
